backend/requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
fonttools==4.58.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
narwhals==1.41.0
numpy==2.2.6
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
wheel==0.45.1
